Created templates for django-registration.
Added django-registration to installed apps. Created a login redirect url to the main template page in the settings.py. Manually edited files in django-registration; the package is no longer maintained and does not work with Django 1.7. Created templates for login, logout, registration complete and registration form. Edited the length of the columns for user input and added glyphicons inside of the user input fields. Added code in the login and registration template to raise an error is user information is invalid.