Added a view, template and url for the PeopleView.
Created the url and test for the PeopleView. Added view for the people. The view included a function that would take a list of letters and collect the last names of the people in correlation to the letter. The letters where then placed in a dictionary to be used in the template. This is necessary because of our large amount of data. Created the template that included buttons that expanded to show the list of people in correlation to the letters along with their name, birth and death date. The name was made into a link that redirects the user to the detail view for the person’s name. Updated the navigation bar in the base.html to include the search person’s last name by category.