Edit MarkerDetailView.
Added fixture_media to installed apps. Note two of the original files needed to be altered to work with Python 3. Enabled the server to host the Media files.  Updated fixture to include image.  Changed MarkerDetailView from DetailView to TemplateView.  Created a function in MarkerDetailView to resize images in database while preserving the images aspect ratio.  Also included a statement to display detailed information about marker if image did not exist in the markers_detail template.  Created model test for MarkerImage.