Altered the way views pass data to templates and templates.
Added context data to  PersonDetailView. Context data allows the image of the person's grave marker to display on the person_detail template. Altered context data on the MarkerDetailView to include person. Added links to the names of the people who are listed on the grave marker to marker_detail template.