Altered templates, initial_data, and views.
Added columns and rows to the the person detail template and the markers detail template so information about the person and marker could be displayed to the left of the image. Changed the font to matching fonts in all templates. Changed the size of the resized image  in the person detail view and the markers detail view.