Added HttpRequest.META to views and altered templates.
Fixed the return to previous page links on the person_detail template and the makers_detail template to return the user to the proper link based on the original page the user was viewing. Added url template tags to the return href in both templates. Altered the markerDetailView to include Http_REFERER.