Altered model and updated initial data fixture
Changed the maximum length of characters on the marker model family name field. The field was not large enough to hold some of the markers added to the database. Also created an additional initial data fixture to only include the cemetery name and description. We wrote script to load data from shapefiles but data included in both the shapefile and the initial data fixture were repeated twice in the database.