Updated project for two additional maps
Added SymbologyMapView and created a function, when ‘see symbology map’ link is clicked on the symbology template(displayed above markerids of each symbol, queries the database and sends the markers matching the selected symbology to the geojson data template. The selected markers then display on the symbology map. Created view, url, and template for map of persons. Template is similar to marker map template aside from links. Altered namespace for marker map in url, templates(includes base and carousel), model and test. The term map was to generic  making it difficult to determine what the term map stood for. Added link on the nav bar in base.html template for the person map but not for the symbology map. Symbology map can only be assessed from the symbology template.  Also added test for new urls.