Added required files for API.
Installed django-restframework and added it to settings file and requirements. Created serializers.py and added Meta for Person and Markers. Also added fields we wanted to include in our api. Added json_views to set up the GET request. At this time, we do not want to allow the API to create, update or delete so we excluded those mixins. Created a separate urls for the API. Started working on test, but it requires some additional work. The Approximate Date Field has a different date format and caused the test to fail.   Url test for the api are, however, up to date. Updated geojson and altered symbology map to include new syntax for the upgrade.