Added  Items for WWI Draft Cards.
Created model with filefield for pdf file and foreign keyed to person. Added test for model. Registered new model with admin.py. Created two views and templates. The first view and template shows a list of all eligible draftees based on birth dates and includes link to a pdf if draft card exists in the database. The second view uses the api to display points for the draftees on a map. Added urls for both templates and created test s. Altered PersonDetailView and person detail template to include links to pdf if it exists.  Altered person filter in json_views and person serializer to include id. Without the id, the api id field displayed blank and would not allow a clickable link on map to redirect to PersonDetailView. Updated base.html to include links to list of draftees and map of draftees. Created new fixture to hold draft card pdf. Updated initial_data fixture to include recoded for person who had a draft card pdf.